Update package.json, add typecheck, PR comments
