+++ .github/workflows/android-device-test.yml
@@ -131,7 +131,7 @@ jobs:
       - name: Upload benchmark results to S3
         run: |
           npm install --ignore-scripts
-          node .github/workflows/android-device-test/download-benchmark-results.mjs ${{ steps.aws_device_farm_run.outputs.runArn }}
+          node scripts/aws-device-farm/download-benchmark-results.mjs ${{ steps.aws_device_farm_run.outputs.runArn }}
 
       - uses: LouisBrunner/checks-action@v2.0.0
         if: always() && (env.run_device_test == 'true' || matrix.test.name == 'Android Render Tests')
