+++ package.json
@@ -27,10 +27,13 @@
     "npm-run-all": "^4.1.5"
   },
   "devDependencies": {
+    "@aws-sdk/client-device-farm": "^3.556.0",
+    "@aws-sdk/client-s3": "^3.556.0",
     "@mapbox/flow-remove-types": "^2.0.0",
     "@mapbox/mvt-fixtures": "3.10.0",
     "@octokit/plugin-retry": "^7.1.0",
     "@octokit/rest": "^20.1.0",
+    "@types/node": "^20.12.7",
     "argparse": "^2.0.1",
     "csscolorparser": "~1.0.3",
     "d3-queue": "3.0.7",
@@ -51,6 +54,7 @@
     "shuffle-seed": "1.1.6",
     "st": "3.0.0",
     "tape": "^5.7.5",
+    "typescript": "^5.4.5",
     "xcode": "^3.0.1"
   },
   "engines": {
