+++ scripts/jsconfig.json
@@ -1,0 +1,11 @@
+{
+  "compilerOptions": {
+    "strict": true,
+    "allowJs": true,
+    "checkJs": true,
+    "module": "NodeNext"
+  },
+  "include": [
+    "**/*.mjs"
+  ]
+}
